Add files via upload
The game starts with two tiles, each displaying either "2" or "4," randomly placed on the 4x4 grid.
Players can swipe in four directions: up, down, left, or right. All tiles on the grid will move in the chosen direction until they hit the edge or another tile.
When two tiles with the same number collide while swiping, they merge into a new tile with a value equal to the sum of the original tiles.
For instance, merging two "2" tiles creates a "4" tile, and combining two "4" tiles results in an "8" tile, and so on.
After every successful swipe, a new tile (either "2" or "4") appears on the grid at an empty spot.
The game ends when the grid is full, and there are no more possible moves, i.e., no empty spots and no adjacent tiles with matching numbers.
The player's goal is to combine tiles and achieve the "2048" tile. However, players can continue playing even after reaching 2048 to achieve higher numbers and aim for a higher score.
Strategies and Tips:

To progress effectively, players must plan their moves carefully. A wrong move may lead to quickly filling up the grid and blocking potential matches.
Players should focus on keeping the largest numbers in one corner or along one edge of the grid to minimize the risk of getting trapped among smaller tiles.
It is vital to maintain open spaces on the grid for future moves, so it's crucial not to let the largest numbers get isolated from potential matches.
Players must also be cautious about creating a pattern that constantly repeats itself, as it may restrict their ability to merge tiles effectively.
Scoring:

Each time two tiles combine, the player earns points equal to the value of the new tile.
For instance, merging two "16" tiles creates a "32" tile and awards 32 points, and so on.
The game keeps track of the player's highest score achieved during the current session.
Popularity and Legacy:
2048 quickly gained popularity and became a viral sensation due to its simple yet challenging gameplay and the desire to achieve the coveted "2048" tile. Initially developed as an open-source project, the game has inspired numerous variations and adaptations across various platforms and devices.

Conclusion:
2048 is a timeless classic in the world of mobile gaming, loved by puzzle enthusiasts of all ages. With its addictive nature and the quest to reach the magical "2048" tile, the game offers endless hours of entertainment and remains a testament to the creativity and ingenuity of its creator. Whether played casually or competitively, 2048 continues to hold its place as one of the most beloved and celebrated puzzle games of all time.